requirements.txt
playwright==1.36.0
pandas==2.0.3
pydantic==2.10.6
pydantic-settings==2.8.1
tqdm==4.65.0
python-dotenv==1.0.0